+++ tests/test_playbook_grapher.py
@@ -30,7 +30,7 @@ def run_grapher(playbook_file: str, output_filename: str = None, additional_args
 
     if "--open-protocol-handler" not in additional_args:
         additional_args.insert(0, "--open-protocol-handler")
-        additional_args.insert(1, os.environ.get("TEST_OPEN_PROTOCOL", "browser"))
+        additional_args.insert(1, "vscode")
 
     playbook_path = os.path.join(FIXTURES_DIR, playbook_file)
     args = [__prog__]